--- seminar0/src/main/kotlin/com.wafflestudio.seminar/survey/api/SeminarController.kt
@@ -1,32 +1,0 @@
-package com.wafflestudio.seminar.survey.api
-
-import com.wafflestudio.seminar.survey.service.SeminarService
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
-import org.springframework.web.bind.annotation.RequestParam
-import org.springframework.web.bind.annotation.RestController
-
-@RestController
-class SeminarController(
-    private val service: SeminarService
-) {
-    @GetMapping("/os/{osId}")
-    fun getOs(
-        @PathVariable osId: Long,
-    ) = service.os(osId)
-
-    @GetMapping("/os")
-    fun getOs(
-        @RequestParam name: String
-    ) = service.os(name)
-
-    @GetMapping("/survey")
-    fun getSurveyList() =
-        service.surveyResponseList()
-
-    @GetMapping("/survey/{surveyId}")
-    fun getSurvey(
-        @PathVariable surveyId: Long,
-    ) = service.surveyResponse(surveyId)
-
-}
