+++ seminar0/src/main/kotlin/com.wafflestudio.seminar/survey/database/OsRepositoryImpl.kt
@@ -8,6 +8,15 @@ class OsRepositoryImpl(private val memoryDB : MemoryDB): OsRepository {
     override fun findAll(): List<OperatingSystem> {
         return memoryDB.getOperatingSystems()
     }
+    override fun findByName(name:String):OperatingSystem? {
+            val list : List<OperatingSystem> =memoryDB.getOperatingSystems()
+            list.forEach{
+                if(it.osName == name){
+                    return it
+                }
+            }
+            return null
+        }
 
     override fun findById(id: Long): OperatingSystem? {
         val list : List<OperatingSystem> =memoryDB.getOperatingSystems()
