+++ seminar0/src/main/kotlin/com.wafflestudio.seminar/survey/service/SeminarService.kt
@@ -4,33 +4,35 @@ import com.wafflestudio.seminar.survey.database.OsRepository
 import com.wafflestudio.seminar.survey.database.SurveyResponseRepository
 import com.wafflestudio.seminar.survey.domain.OperatingSystem
 import com.wafflestudio.seminar.survey.domain.SurveyResponse
-import org.springframework.stereotype.Service
+import com.wafflestudio.seminar.survey.exception.SeminarException
+import org.springframework.http.HttpStatus
+import org.springframework.stereotype.Component
 
-interface SeminarService {
-    fun os(name: String): OperatingSystem
-    fun os(id: Long): OperatingSystem
-    fun surveyResponseList(): List<SurveyResponse>
-    fun surveyResponse(id: Long): SurveyResponse
-}
-
-@Service
-class SeminarServiceImpl(
-    private val surveyResponseRepository: SurveyResponseRepository,
-    private val osRepository: OsRepository,
-) : SeminarService {
-    override fun os(id: Long): OperatingSystem {
-        return osRepository.findById(id)
+@Component
+class Service(private val osRepository : OsRepository, private val surveyResponseRepository : SurveyResponseRepository) {
+    fun findByOsId(id: Long): OperatingSystem {
+        if(osRepository.findById(id)!=null){return osRepository.findById(id)!!}
+        throw SeminarException("Cannot find the os with the id of $id", HttpStatus.BAD_REQUEST)
     }
 
-    override fun os(name: String): OperatingSystem {
-        return osRepository.findByName(name)
+    fun findByOsName(name: String): OperatingSystem {
+        var totalList: MutableList<OperatingSystem> = osRepository.findAll().toMutableList()
+        totalList.forEach{
+            if(it.osName.equals(name)){
+                return it
+            }
+        }
+        throw SeminarException("Cannot find the os with the name of $name", HttpStatus.BAD_REQUEST)
     }
 
-    override fun surveyResponseList(): List<SurveyResponse> {
+    fun findAllSurveyResponse(): List<SurveyResponse> {
         return surveyResponseRepository.findAll()
     }
 
-    override fun surveyResponse(id: Long): SurveyResponse {
-        return surveyResponseRepository.findById(id)
+    fun findBySurveyResponseId(id: Long): SurveyResponse {
+        if(surveyResponseRepository.findById(id)!=null){
+            return surveyResponseRepository.findById(id)!!
+        }
+        throw SeminarException("Cannot find the survey response with the id of $id", HttpStatus.BAD_REQUEST)
     }
 }
