--- seminar0/src/main/kotlin/com/wafflestudio/seminar/survey/database/OsRepository.kt
@@ -1,8 +1,0 @@
-package com.wafflestudio.seminar.survey.database
-
-import com.wafflestudio.seminar.survey.domain.OperatingSystem
-
-interface OsRepository {
-    fun findAll(): List<OperatingSystem>
-    fun findById(id: Long): OperatingSystem?
-}
