--- seminar0/src/main/kotlin/com/wafflestudio/seminar/survey/database/SurveyResponseRepository.kt
@@ -1,8 +1,0 @@
-package com.wafflestudio.seminar.survey.database
-
-import com.wafflestudio.seminar.survey.domain.SurveyResponse
-
-interface SurveyResponseRepository {
-    fun findAll(): List<SurveyResponse>
-    fun findById(id: Long): SurveyResponse?
-}
