--- seminar0/src/main/kotlin/com/wafflestudio/seminar/survey/exception/SeminarException.kt
@@ -1,7 +1,0 @@
-package com.wafflestudio.seminar.survey.exception
-
-import org.springframework.http.HttpStatus
-
-class SeminarException(val errorMessage: String,val httpstatus:HttpStatus) : RuntimeException() {
-    
-}
